requirement.txt
torch
torchvision
ujson
pyyaml